MB-22008: Introduce CV Jenkinsfile
Allows the use of Multibranch Pipeline Jobs in Jenkins to centralise
configuration of the various jobs. Many of the existing jobs simply call
single-project-gerrit with different env variables to configure it.

This commit introduces a Jenkinsfile reimplementation of single-project-gerrit
and single-project-clang-static-analyzer, and a set of config files specifying
the env variables for a given job.

Change-Id: I92bb0db6ac2b526857e5191a62f368d8e9284026